requirements.txt
langchain-google-genai
google-generativeai
flask>=2.3.0 
requests>=2.31.0
python-dotenv>=1.0.0
langchain-pinecone>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.8.0
langchain-openai>=0.1.0
openai>=1.0.0
langchain-core>=0.2.0
langchain-community>=0.2.0
gunicorn>=21.2.0
pytest>=7.4.0
black>=23.0.0
pinecone-client==3.0.0

